+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Show number of occurrences when a response isn't a 200 response in responseOk advice.
 - Rewrite fastRender advice and made it include cssInHeadDomain (removed) to make it easier to rewrite the advice later and act on both what's in the DOM and the HTML. We need to add more tests before this can be released.
 - We know also can (and do) test our test cases against a H2 server (that was easy).
+- Drop the cssInHead advice, it is now a anti pattern. https://jakearchibald.com/2016/link-in-body/
 
 ## 0.14.0 2016-03-24
 ### Added